Turn ebpf c code into an embed rather than an inline string
